DisplayToPlayer added and relevant interfaces along with doc improvement
Added the Class DisplayToPlayer, along with a package containing the interfaces for displayFunctionality. 
Additionally i added doc comments for some methods, primarily in the three Command classes.
The initial work on Display addresses the CommandWords class for now, but the next step is tackling Game texts.

So far there is still not a deep clear cut user guide on how to use the Class, but it will come in later commits.